✨ implement CreateNewStoryPostTimelineItem components
