+++ src/components/domains/storyPost/CreateNewStoryPostTimelineItem/index.ts
@@ -1,0 +1,1 @@
+export { CreateNewStoryPostTimelineItem } from './CreateNewStoryPostTimelineItem';
